requirements.txt
cryptography.fernet
getpass
tabulate
sys
csv
re